Refactor: EvoModel to own module
Refactored to move the EvolutionaryModel trait to its own module so that
evolutionary models are not just substitution models.
