+++ phylo/src/substitution_models/mod.rs
@@ -1,14 +1,14 @@
-use crate::{
-    f64_h,
-    likelihood::{EvolutionaryModelInfo, LikelihoodCostFunction},
-    phylo_info::PhyloInfo,
-    tree::NodeIdx,
-    Result,
-};
-use nalgebra::{Const, DMatrix, DVector, DimMin, SMatrix, SVector};
-use ordered_float::OrderedFloat;
 use std::collections::HashMap;
 use std::ops::Mul;
+
+use nalgebra::{Const, DMatrix, DimMin, SMatrix, SVector};
+use ordered_float::OrderedFloat;
+
+use crate::evolutionary_models::EvolutionaryModel;
+use crate::f64_h;
+use crate::likelihood::{EvolutionaryModelInfo, LikelihoodCostFunction};
+use crate::phylo_info::PhyloInfo;
+use crate::tree::NodeIdx;
 
 pub mod dna_models;
 pub mod protein_models;
@@ -21,24 +21,6 @@ pub struct SubstitutionModel<const N: usize> {
     index: [i32; 255],
     pub q: SubstMatrix<N>,
     pub pi: FreqVector<N>,
-}
-
-pub trait EvolutionaryModel<const N: usize> {
-    fn new(model_name: &str, model_params: &[f64]) -> Result<Self>
-    where
-        Self: std::marker::Sized;
-    fn get_p(&self, time: f64) -> SubstMatrix<N>;
-    fn get_rate(&self, i: u8, j: u8) -> f64;
-    fn generate_scorings(
-        &self,
-        times: &[f64],
-        zero_diag: bool,
-        rounded: bool,
-    ) -> HashMap<OrderedFloat<f64>, (SubstMatrix<N>, f64)>;
-    fn normalise(&mut self);
-    fn get_scoring_matrix(&self, time: f64, rounded: bool) -> (SubstMatrix<N>, f64);
-    fn get_stationary_distribution(&self) -> &FreqVector<N>;
-    fn get_char_probability(&self, char: u8) -> DVector<f64>;
 }
 
 impl<const N: usize> SubstitutionModel<N>
